change maxAmount to U256g

+++ solidity/src/libraries/Forwarder.sol
@@ -17,7 +17,7 @@ contract Forwarder is Library {
      */
     struct ForwardingConfig {
         address tokenAddress;
-        uint128 maxAmount;
+        uint256 maxAmount;
     }
 
     /**
